# apps/backend/internal/adapters/postgres/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlc

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Account struct {
	ID         pgtype.UUID
	Email      string
	State      string
	Username   string
	CreatedAt  pgtype.Timestamptz
	VerifiedAt pgtype.Timestamptz
	ExpiresAt  pgtype.Timestamptz
}

type DraftTeam struct {
	ID             pgtype.UUID
	DraftID        pgtype.UUID
	Name           string
	NameNormalized string
	Position       int32
}

type EmailVerificationToken struct {
	ID            pgtype.UUID
	AccountID     pgtype.UUID
	TokenHash     []byte
	CreatedAt     pgtype.Timestamptz
	ExpiresAt     pgtype.Timestamptz
	ConsumedAt    pgtype.Timestamptz
	InvalidatedAt pgtype.Timestamptz
}

type ExternalIdentity struct {
	ID        pgtype.UUID
	AccountID pgtype.UUID
	Provider  string
	Issuer    string
	Subject   string
	CreatedAt pgtype.Timestamptz
}

type FederatedLoginChallenge struct {
	ID         pgtype.UUID
	Provider   string
	NonceHash  []byte
	CreatedAt  pgtype.Timestamptz
	ExpiresAt  pgtype.Timestamptz
	ConsumedAt pgtype.Timestamptz
}

type IdentityLinkAttempt struct {
	ID                 pgtype.UUID
	CandidateAccountID pgtype.UUID
	Provider           string
	Issuer             string
	Subject            string
	TokenHash          []byte
	CreatedAt          pgtype.Timestamptz
	ExpiresAt          pgtype.Timestamptz
	ConsumedAt         pgtype.Timestamptz
	InvalidatedAt      pgtype.Timestamptz
}

type League struct {
	ID                 pgtype.UUID
	OrganizerAccountID pgtype.UUID
	Name               string
	Sport              string
	Format             string
	State              string
	RoundRobinLegs     int16
	PointsForWin       int16
	PointsForDraw      int16
	PointsForLoss      int16
	CreatedAt          pgtype.Timestamptz
	PublishedAt        pgtype.Timestamptz
}

type LeagueAdministrator struct {
	LeagueID   pgtype.UUID
	AccountID  pgtype.UUID
	AssignedAt pgtype.Timestamptz
}

type LeagueDraft struct {
	ID        pgtype.UUID
	AccountID pgtype.UUID
	Name      string
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
	ExpiresAt pgtype.Timestamptz
}

type LeagueFollower struct {
	LeagueID   pgtype.UUID
	AccountID  pgtype.UUID
	FollowedAt pgtype.Timestamptz
}

type LeagueTeam struct {
	ID             pgtype.UUID
	LeagueID       pgtype.UUID
	Name           string
	NameNormalized string
	Position       int32
}

type LocalCredential struct {
	AccountID    pgtype.UUID
	PasswordHash string
	CreatedAt    pgtype.Timestamptz
	UpdatedAt    pgtype.Timestamptz
}

type Match struct {
	ID          pgtype.UUID
	LeagueID    pgtype.UUID
	RoundNumber int32
	Sequence    int32
	HomeTeamID  pgtype.UUID
	AwayTeamID  pgtype.UUID
	State       string
}

type Session struct {
	ID                pgtype.UUID
	AccountID         pgtype.UUID
	TokenHash         []byte
	CreatedAt         pgtype.Timestamptz
	LastSeenAt        pgtype.Timestamptz
	IdleExpiresAt     pgtype.Timestamptz
	AbsoluteExpiresAt pgtype.Timestamptz
	RevokedAt         pgtype.Timestamptz
}